+++ test/controllers/internal_controller_test.rb
@@ -14,9 +14,33 @@ class InternalControllerTest < ActionDispatch::IntegrationTest
     assert_response :success
   end
 
+  test 'should get research' do
+    sign_in_as(@regular_user, 'password')
+    get research_path
+    assert_response :success
+  end
+
   test 'should get settings' do
     sign_in_as(@regular_user, 'password')
-    get dashboard_path
+    get settings_path
+    assert_redirected_to settings_profile_path
+  end
+
+  test 'should get settings account' do
+    sign_in_as(@regular_user, 'password')
+    get settings_account_path
+    assert_response :success
+  end
+
+  test 'should get settings consents' do
+    sign_in_as(@regular_user, 'password')
+    get settings_consents_path
+    assert_response :success
+  end
+
+  test 'should get settings emails' do
+    sign_in_as(@regular_user, 'password')
+    get settings_emails_path
     assert_response :success
   end
 end